+++ lib/attr_json/config.rb
@@ -3,10 +3,20 @@ module AttrJson
   # and rails class_attribute. Instead, you set to new Config object
   # changed with {#merge}.
   class Config
-    RECORD_ALLOWED_KEYS = %i{default_container_attribute default_accepts_nested_attributes}
-    MODEL_ALLOWED_KEYS = %i{unknown_key bad_cast}
+    RECORD_ALLOWED_KEYS = %i{
+      default_container_attribute
+      default_rails_attribute
+      default_accepts_nested_attributes
+    }
+
+    MODEL_ALLOWED_KEYS = %i{
+      unknown_key
+      bad_cast
+    }
+
     DEFAULTS = {
       default_container_attribute: "json_attributes",
+      default_rails_attribute: false,
       unknown_key: :raise
     }