+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/jrochkind/attr_json/compare/v1.1.0...HEAD)
+## [Unreleased](https://github.com/jrochkind/attr_json/compare/v1.2.0...HEAD)
+
+
+
+## [1.2.0](https://github.com/jrochkind/attr_json/compare/v1.1.0...v1.2.0)
 
 ### Added